+++ src/main/java/com/projects/oleksii/leheza/cashtruck/dto/view/ClientStatisticDto.java
@@ -3,7 +3,6 @@ package com.projects.oleksii.leheza.cashtruck.dto.view;
 import com.projects.oleksii.leheza.cashtruck.domain.Transaction;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
@@ -16,13 +15,13 @@ public class ClientStatisticDto {
 
     private List<Transaction> expenses;
     private List<Transaction> incomes;
-    private BigDecimal totalIncomeSum;
-    private BigDecimal totalExpenseSum;
-    private BigDecimal totalBalance;
-    private BigDecimal lastYearExpense;
-    private BigDecimal lastMonthExpense;
-    private BigDecimal lastWeekExpense;
-    private BigDecimal lastYearIncome;
-    private BigDecimal lastMonthIncome;
-    private BigDecimal lastWeekIncome;
+    private long totalIncomeSum;
+    private long totalExpenseSum;
+    private long totalBalance;
+    private long lastYearExpense;
+    private long lastMonthExpense;
+    private long lastWeekExpense;
+    private long lastYearIncome;
+    private long lastMonthIncome;
+    private long lastWeekIncome;
 }
