+++ src/main/java/com/projects/oleksii/leheza/cashtruck/dto/view/UserDto.java
@@ -4,7 +4,6 @@ import com.projects.oleksii.leheza.cashtruck.domain.BankCard;
 import com.projects.oleksii.leheza.cashtruck.domain.Transaction;
 import lombok.*;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
@@ -29,7 +28,7 @@ public class UserDto {
     private String role;
     private String status;
     private String subscription;
-    private BigDecimal subscriptionPrice;
+    private int subscriptionPrice;
     private Date subscriptionFinishDate;
     private List<Transaction> incomes;
     private List<Transaction> expenses;
